added only number validation and round off bank balance to 2dp for update cash balance

+++ pmsapp/src/main/java/com/pms/pmsapp/performance/controller/PerformanceController.java
@@ -110,7 +110,7 @@ public class PerformanceController {
 	}
 
 	@RequestMapping(value = "/performance/updateCashBal", method = RequestMethod.POST)
-	public PerformanceForm updateCashBal(@RequestBody Map<String, Integer> jsonMap, Authentication authentication) {
+	public PerformanceForm updateCashBal(@RequestBody Map<String, BigDecimal> jsonMap, Authentication authentication) {
 		log.info("performance updateCashBal in Controller");
 
 		String username = authentication.getName();
