+++ pmsapp/src/main/java/com/pms/pmsapp/performance/service/PerformanceServiceImpl.java
@@ -41,7 +41,7 @@ public class PerformanceServiceImpl implements PerformanceService {
 	}
 
 	@Override
-	public String updateCashBal(int newCashBal, String username) throws Exception {
+	public String updateCashBal(BigDecimal newCashBal, String username) throws Exception {
 		return performanceDao.updateCashBal(newCashBal, username);
 	}
 }
